Implement Sound Object Load
