+++ src/OpenLoco/src/Audio/Audio.h
@@ -79,6 +79,19 @@ namespace OpenLoco::Audio
         uint16_t endYear;
     };
 
+#pragma pack(push, 1)
+    struct WAVEFORMATEX
+    {
+        int16_t wFormatTag;
+        int16_t nChannels;
+        int32_t nSamplesPerSec;
+        int32_t nAvgBytesPerSec;
+        int16_t nBlockAlign;
+        int16_t wBitsPerSample;
+        int16_t cbSize;
+    };
+#pragma pack(pop)
+
     void initialiseDSound();
     void disposeDSound();
     void close();
