+++ src/OpenLoco/src/Objects/SoundObject.cpp
@@ -1,16 +1,30 @@
 #include "SoundObject.h"
 #include "Interop/Interop.hpp"
+#include "ObjectStringTable.h"
+#include <cassert>
 
 namespace OpenLoco
 {
     // 0x0048AFAF
     void SoundObject::load(const LoadedObjectHandle& handle, stdx::span<const std::byte> objData, ObjectManager::DependentObjects*)
     {
-        Interop::registers regs;
-        regs.esi = Interop::X86Pointer(this);
-        regs.ebx = handle.id;
-        regs.ecx = enumValue(handle.type);
-        Interop::call(0x0048AFAF, regs);
+        auto remainingData = objData.subspan(sizeof(SoundObject));
+
+        auto strRes = ObjectManager::loadStringTable(remainingData, handle, 0);
+        name = strRes.str;
+        remainingData = remainingData.subspan(strRes.tableLength);
+
+        // Unknown structure size 0x10
+        uint32_t numUnkStructs = *reinterpret_cast<const uint32_t*>(remainingData.data());
+        remainingData = remainingData.subspan(sizeof(uint32_t));
+        uint32_t pcmDataLength = *reinterpret_cast<const uint32_t*>(remainingData.data());
+        remainingData = remainingData.subspan(sizeof(uint32_t));
+        remainingData = remainingData.subspan(numUnkStructs * 16);
+
+        // Main pcm data structure
+        data = reinterpret_cast<const SoundObjectData*>(remainingData.data());
+
+        assert(remainingData.size() == pcmDataLength);
     }
 
     // 0x0048AFE1
