+++ src/OpenLoco/src/Objects/SoundObject.h
@@ -1,5 +1,6 @@
 #pragma once
 
+#include "Audio/Audio.h"
 #include "Object.h"
 #include "Types.hpp"
 #include <OpenLoco/Core/Span.hpp>
@@ -15,12 +16,26 @@ namespace OpenLoco
         constexpr SoundObjectId_t null = 0xFF;
     }
 #pragma pack(push, 1)
+
+    struct SoundObjectData
+    {
+        int32_t var_00;
+        int32_t offset;
+        uint32_t length;
+        Audio::WAVEFORMATEX pcmHeader;
+
+        const void* pcm() const
+        {
+            return (void*)((uintptr_t)this + sizeof(SoundObjectData));
+        }
+    };
+
     struct SoundObject
     {
         static constexpr auto kObjectType = ObjectType::sound;
 
         string_id name;
-        void* data;
+        const SoundObjectData* data;
         uint8_t var_06;
         uint8_t pad_07;
         uint32_t volume; // 0x08
